TheTaskDelegatorBot/requirements.txt
aiogram==3.0.0b7
SQLAlchemy==2.0.21
python-dotenv==1.0.0
matplotlib==3.7.2
numpy==1.24.3
Pillow==10.1.0
requests==2.31.0